refactor: Substitute peter-evans/create-pull-request with gh cli
Not replicating the `delete-branch` functionality of
`peter-evans/create-pull-request`, that, upon next run and before
creating the PR, would delete the branch to make sure the branch was
pristine. This is because we now have the `policy-catalog` repo
configured as instructed by `peter-evans/create-pull-request` with
"Automatically delete head branches feature" in the repository settings.

Signed-off-by: Víctor Cuadrado Juan <vcuadradojuan@suse.de>